implements tests for granular and regd extra controller
